Drop USD API and just expose a managed singleton
In theory putting things into USD properties
should make sure that anyone can consume them. In
practice listening to USD properties is a pain
and I never managed to do it robustly. Having
a single global instance which we expect
Python extension consumers to grab creates opportunities
for different consumers to mess up the device object,
but at least it's very easy to work with.
This means that the device is decoupled from the
stage now, so you can engage the device just once
for the life of the Sim process.
Allows us to drop PhysX and Isaac dependencies,
which were weird to begin with.